rename: remove _new_ prefix from test files
