+++ python/packages/redis/tests/test_providers.py
